Allows more flexible scanning
This patch uses the BleakScanner interface to simplify the scanning as
well as provide more means of connecting. With an address, a BLEDevice,
or with a user provided scanner.

This change will allow me to implement a Home Assistant integration
using [Bluetooth
discovery](https://developers.home-assistant.io/docs/network_discovery/#best-practices-for-library-authors).

Other changes: `black`